separate out client and server code into their own servers, to make it easier to see what code's client-side and what's server-side 🐿 v2.12.5
