+++ client/components/bottom/app-promo-mobile/main.scss
