+++ client/components/bottom/cookie-consent/main.js
