+++ client/components/bottom/cookie-consent/main.scss
