+++ client/components/bottom/fast-ft/main.scss
