+++ client/components/bottom/ft-weekend-promo/main.scss
