+++ client/components/bottom/markets-data/main.scss
