+++ client/components/bottom/onboarding-app/main.js
