+++ client/components/bottom/onboarding-app/main.scss
