+++ client/components/bottom/swg-entitlements-prompt/main.js
