+++ client/components/bottom/team-trial/main.scss
