+++ client/components/bottom/tech-scroll-asia/main.scss
