+++ client/components/top/onboarding-region/main.js
