+++ client/next-messaging-guru-client.js
