+++ client/top-slot.js
