+++ main-server.js
@@ -1,0 +1,3 @@
+module.exports = {
+	presenter: require('./server/nMessagingPresenter')
+};
