--- main.js
@@ -1,3 +1,0 @@
-module.exports = {
-	presenter: require('./src/handlebars-helpers/nMessagingPresenter')
-};
