+++ main.scss
@@ -1,13 +1,13 @@
-@import './src/components/bottom/onboarding-app/main';
-@import './src/components/top/anon-subscribe-now/main';
-@import './src/components/bottom/team-trial/main';
-@import './src/components/bottom/b2b-trial-myft/main';
-@import './src/components/bottom/b2b-trial-mobile/main';
-@import './src/components/bottom/tech-scroll-asia/main';
-@import './src/components/bottom/app-promo-mobile/main';
-@import './src/components/bottom/newsletter-promo/main';
-@import './src/components/bottom/ft-weekend-promo/main';
-@import './src/components/bottom/my-ft-feedpage-overview/main';
-@import './src/components/bottom/fast-ft/main';
-@import './src/components/bottom/markets-data/main';
-@import './src/components/top/nbe-auto-sub/main';
+@import './client/components/bottom/onboarding-app/main';
+@import './client/components/top/anon-subscribe-now/main';
+@import './client/components/bottom/team-trial/main';
+@import './client/components/bottom/b2b-trial-myft/main';
+@import './client/components/bottom/b2b-trial-mobile/main';
+@import './client/components/bottom/tech-scroll-asia/main';
+@import './client/components/bottom/app-promo-mobile/main';
+@import './client/components/bottom/newsletter-promo/main';
+@import './client/components/bottom/ft-weekend-promo/main';
+@import './client/components/bottom/my-ft-feedpage-overview/main';
+@import './client/components/bottom/fast-ft/main';
+@import './client/components/bottom/markets-data/main';
+@import './client/components/top/nbe-auto-sub/main';
