+++ test/server/nMessagingPresenter.test.js
@@ -1,5 +1,4 @@
-const MESSAGING_PRESENTER = '../../src/handlebars-helpers/nMessagingPresenter';
-const SLOT_PRESENTER = '../presenters/slot-presenter';
+const MESSAGING_PRESENTER = '../../server/nMessagingPresenter';
 
 const { expect } = require('chai');
 const sinon = require('sinon');
@@ -39,14 +38,6 @@ describe('nMessagingPresenter', () => {
 		expect(options.fn.callCount).to.equal(1);
 	});
 
-	it('should initalising the SlotPresenter if options has data', () => {
-		const proxyNMessagingPresenter = proxyquire(MESSAGING_PRESENTER, {
-			[SLOT_PRESENTER]: stub
-		});
-		proxyNMessagingPresenter(context, options);
-		expect(stub.callCount).to.equal(1);
-	});
-
 	it('should not run the options.fn method if options has no data', () => {
 		sinon.spy(options, 'fn');
 		delete options.data;
@@ -54,12 +45,4 @@ describe('nMessagingPresenter', () => {
 		expect(options.fn.callCount).to.equal(0);
 	});
 
-	it('should not initalise the SlotPresent if options has no data', () => {
-		const proxyNMessagingPresenter = proxyquire(MESSAGING_PRESENTER, {
-			[SLOT_PRESENTER]: stub
-		});
-		delete options.data;
-		proxyNMessagingPresenter(context, options);
-		expect(stub.callCount).to.equal(0);
-	});
 });
